// panels/src/app/models/ResponseDataGetAll.ts
interface Sort {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
}
interface Pageable {
    offset: number;
    pageNumber: number;
    pageSize: number;
    paged: boolean;
    sort: Sort;
    unpaged: boolean;
}
export interface ResponseDataGetAll<T> {
    content: T[];
    empty: boolean;
    first: boolean
    last: boolean;
    number: number;
    numberOfElements: number;
    pageable: Pageable;
    size: number;
    sort: Sort;
    totalElements: number;
    totalPages: number;

}